# SpringBoot/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"Spring Boot Api Example","description":"Generic Spring Boot Examples","contact":{"name":"Tobias Emiliano Moreno","email":"tobiasmoreno.tm.21@gmail.com"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.desc@'"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"Create a new Player","description":"Return the player created with te id. If a player exist with the username or email, return 404.Additionally, the email must be valid and the pass must have at least 8 characters","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internet Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"400":{"description":"Username or Email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/login":{"post":{"tags":["login-controller"],"summary":"Login a player in the plataform","description":"Return the player logged in if the credentials are correct.","operationId":"loginPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credential"}}},"required":true},"responses":{"404":{"description":"Credentials Invalid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internet Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/login/v2":{"post":{"tags":["login-controller"],"summary":"Login a player in the plataform","description":"Return the player logged in if the credentials are correct.","operationId":"loginPlayer_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialV2"}}},"required":true},"responses":{"404":{"description":"Credentials Invalid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internet Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Description of the method. Get Player by Id","description":"Description of the method. Return a player by th id. If the player doesn´t exist return 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internet Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}/matches":{"get":{"tags":["player-controller"],"operationId":"getMatchesOfPlayer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Match"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Description of the method","description":"Description of the method. If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internet Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"The player ID","format":"int64","example":1},"userName":{"title":"Player user name","type":"string","description":"The player userName","example":"TobiasExample"},"password":{"title":"Player ID","type":"string","description":"The player Password","example":"Password#03"},"email":{"title":"Player Email","type":"string","description":"The player Email","example":"email@email.com"},"avatar":{"title":"Player Avatar URL","type":"string","description":"The player Avatar","nullable":true,"example":"http://localhost:8080/avatars/myUsername"},"lastLoginDate":{"title":"Player Last Login","type":"string","description":"The player Last Login","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timeStamp":{"type":"string"},"error":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"}}},"Credential":{"type":"object","properties":{"identity":{"oneOf":[{"$ref":"#/components/schemas/EmailIdentity"},{"$ref":"#/components/schemas/UserNameIdentity"}]},"password":{"type":"string"}}},"EmailIdentity":{"required":["email","identity_type"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"email":{"title":"Email to logged in","type":"string","description":"The player Email provided to logged in","example":"email@email.com"}}}]},"Identity":{"required":["identity_type"],"type":"object","properties":{"identity_type":{"title":"Type of identity used to logged in","type":"string","description":"The Type of identity provided to logged in","example":"USERNAME OR EMAIL","enum":["USERNAME","EMAIL"]}},"discriminator":{"propertyName":"identity_type"}},"UserNameIdentity":{"required":["identity_type","user_name"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"user_name":{"title":"userName to logged in","type":"string","description":"The player userName provided to logged in","example":"myUserName"}}}]},"CredentialV2":{"required":["identity","password"],"type":"object","properties":{"password":{"type":"string"},"identity":{"title":"Email or UserName to logged in","type":"string","description":"The player email or username","example":"email@email.com or username"}}},"Game":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"rules":{"type":"string"}}},"Match":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"game":{"$ref":"#/components/schemas/Game"},"player":{"$ref":"#/components/schemas/Player"},"createdDate":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["STARTED","FINISHED","CANCELED"]}}}}}}